DSR-277: rename CardLink to CardAnchor
